+++ rust/src/trade/requests/replace_order.rs
@@ -1,14 +1,11 @@
 use rust_decimal::Decimal;
 use serde::Serialize;
 
-use crate::serde_utils;
-
 /// Options for replace order request
 #[derive(Debug, Serialize, Clone)]
 pub struct ReplaceOrderOptions {
     order_id: String,
-    #[serde(with = "serde_utils::int64_str")]
-    quantity: i64,
+    quantity: Decimal,
     #[serde(skip_serializing_if = "Option::is_none")]
     price: Option<Decimal>,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -26,7 +23,7 @@ pub struct ReplaceOrderOptions {
 impl ReplaceOrderOptions {
     /// Create a new `ReplaceOrderOptions`
     #[inline]
-    pub fn new(order_id: impl Into<String>, quantity: i64) -> Self {
+    pub fn new(order_id: impl Into<String>, quantity: Decimal) -> Self {
         Self {
             order_id: order_id.into(),
             quantity,
